+++ Editor/TexTransGroup/TexTransGroup.cs
@@ -1,13 +1,15 @@
 #if UNITY_EDITOR
 using UnityEngine;
 using System.Collections.Generic;
+using System.Linq;
+using net.rs64.TexTransTool.Utils;
 
 namespace net.rs64.TexTransTool
 {
+    [AddComponentMenu("TexTransTool/TexTransGroup")]
     public class TexTransGroup : AbstractTexTransGroup
     {
-        public List<TextureTransformer> TextureTransformers = new List<TextureTransformer>();
-        public override IEnumerable<TextureTransformer> Targets => TextureTransformers;
+        public override IEnumerable<TextureTransformer> Targets => transform.GetChildren().ConvertAll(x => x.GetComponent<TextureTransformer>()).Where(x => x != null);
     }
 }
 #endif